Only check content inside #adventure-container element
Previously, the whole body was checked by the assertions.
This lead to false positives, most notably the "'medium'" starting
level range, which was reported to be preset but in fact wasn't. By
restricting the assertions to the #adventure-container element, we make
sure to only check places where we actually expect the string to be.