Update to PHP7.4, ElasticSearch 7.6.2, Ubuntu 18.04 (#353)
This PR updates most of our infrastructure to more recent versions.

PHP 7.0 -> 7.4
ElasticSearch 5.5.3 -> 7.6.2
Ubuntu 16.04LTS -> 18.04 LTS
It also:

removes the type parameter from ElasticSearch, which no longer exists
updates a lot of dependencies to newer minor versions for compatibility with PHP7.4
decreases unnecessarily high memory used by ElasticSearch from 2G to 1G
fixes numerours deprecations
properly uses booleans in DQL instead of numbers, to fix compatibility with PostgreSQL (which will be used for pull request previews on Heroku in an upcoming PR)
introduces a patch needed for the EasyAdmin bundle to work with PHP 7.4. We cannot upgrade to the next major version of the bundle until we upgrade Symfony to the next major version
If you want to try this out, you'll have to vagrant destroy your dev environment and re-create it. Gitpod should work fine, if you use the review button below.